+++ RELEASES.md
@@ -17,6 +17,7 @@ Released on TBD (UTC).
 - Improved `WebSocketClient` error handling and added `WebSocketClientError` exception for Python (#1876), thanks @twitu
 - Improved `WebSocketClient.send_text` efficiency (now accepts UTF-8 encoded bytes, rather than a Python string)
 - Refactored `RedisMessageBusDatabase` to tokio tasks
+- Refactored `RedisCacheDatabase` to tokio tasks
 - Upgraded `tokio` crate to v1.40.0
 
 ### Breaking Changes